+++ lib/components/VegaLiteChart.vue
@@ -2,16 +2,9 @@
   <div class="vega-lite-chart no-drag" :class="{ 'overflow-hidden': !showingControls }">
     <!-- Toggle button always visible -->
     <div class="controls-toggle" v-if="showControls">
-      <button
-        @click="toggleControls"
-        class="toggle-controls-btn"
-        :class="{ active: showingControls }"
-        data-testid="toggle-chart-controls-btn"
-      >
-        <i
-          :class="showingControls ? 'mdi mdi-eye-outline' : 'mdi mdi-cog-outline'"
-          class="icon"
-        ></i>
+      <button @click="toggleControls" class="toggle-controls-btn" :class="{ active: showingControls }"
+        data-testid="toggle-chart-controls-btn">
+        <i :class="showingControls ? 'mdi mdi-eye-outline' : 'mdi mdi-cog-outline'" class="icon"></i>
         <span>{{ showingControls ? 'View Chart' : 'Edit Chart' }}</span>
       </button>
     </div>
@@ -19,27 +12,17 @@
     <!-- Content area with conditional rendering -->
     <div class="chart-content-area">
       <!-- Chart visualization area - only show when controls are hidden -->
-      <div
-        v-show="!showingControls"
-        ref="vegaContainer"
-        class="vega-container"
-        data-testid="vega-chart-container"
-      ></div>
+      <div v-show="!showingControls" ref="vegaContainer" class="vega-container" data-testid="vega-chart-container">
+      </div>
 
       <!-- Controls panel - only show when toggled -->
       <div v-if="showingControls" class="chart-controls-panel">
         <div class="inner-padding">
           <div class="control-section">
             <div class="chart-type-icons">
-              <button
-                v-for="type in charts"
-                :key="type.value"
-                @click="updateConfig('chartType', type.value)"
-                class="chart-icon"
-                :class="{ selected: internalConfig.chartType === type.value }"
-                :title="type.label"
-                :data-testid="`chart-type-${type.value}`"
-              >
+              <button v-for="type in charts" :key="type.value" @click="updateConfig('chartType', type.value)"
+                class="chart-icon" :class="{ selected: internalConfig.chartType === type.value }" :title="type.label"
+                :data-testid="`chart-type-${type.value}`">
                 <div class="icon-container">
                   <i :class="type.icon" class="icon"></i>
                 </div>
@@ -50,24 +33,15 @@
           <!-- Group axes controls  -->
           <div class="control-section" v-if="visibleControls.some((c) => c.filterGroup === 'axes')">
             <label class="control-section-label">Axes</label>
-            <div
-              v-for="control in visibleControls.filter((c) => c.filterGroup === 'axes')"
-              :key="control.id"
-              class="control-group no-drag"
-            >
+            <div v-for="control in visibleControls.filter((c) => c.filterGroup === 'axes')" :key="control.id"
+              class="control-group no-drag">
               <label class="chart-label" :for="control.id">{{ control.label }}</label>
-              <select
-                :id="control.id"
-                :value="internalConfig[control.field]"
+              <select :id="control.id" :value="internalConfig[control.field]"
                 @change="updateConfig(control.field, ($event.target as HTMLInputElement).value)"
-                class="form-select no-drag"
-              >
+                class="form-select no-drag">
                 <option v-if="control.allowEmpty" value="">None</option>
-                <option
-                  v-for="column in filteredColumnsInternal(control.columnFilter)"
-                  :key="column.name"
-                  :value="column.name"
-                >
+                <option v-for="column in filteredColumnsInternal(control.columnFilter)" :key="column.name"
+                  :value="column.name">
                   {{ column.name }}{{ column.description ? ` - ${column.description}` : '' }}
                 </option>
               </select>
@@ -75,29 +49,17 @@
           </div>
 
           <!-- Group appearance controls -->
-          <div
-            class="control-section"
-            v-if="visibleControls.some((c) => c.filterGroup === 'appearance')"
-          >
+          <div class="control-section" v-if="visibleControls.some((c) => c.filterGroup === 'appearance')">
             <label class="control-section-label">Appearance</label>
-            <div
-              v-for="control in visibleControls.filter((c) => c.filterGroup === 'appearance')"
-              :key="control.id"
-              class="control-group no-drag"
-            >
+            <div v-for="control in visibleControls.filter((c) => c.filterGroup === 'appearance')" :key="control.id"
+              class="control-group no-drag">
               <label class="chart-label" :for="control.id">{{ control.label }}</label>
-              <select
-                :id="control.id"
-                :value="internalConfig[control.field]"
+              <select :id="control.id" :value="internalConfig[control.field]"
                 @change="updateConfig(control.field, ($event.target as HTMLInputElement).value)"
-                class="form-select no-drag"
-              >
+                class="form-select no-drag">
                 <option v-if="control.allowEmpty" value="">None</option>
-                <option
-                  v-for="column in filteredColumnsInternal(control.columnFilter)"
-                  :key="column.name"
-                  :value="column.name"
-                >
+                <option v-for="column in filteredColumnsInternal(control.columnFilter)" :key="column.name"
+                  :value="column.name">
                   {{ column.name }}{{ column.description ? ` - ${column.description}` : '' }}
                 </option>
               </select>
@@ -105,46 +67,29 @@
           </div>
 
           <!-- Group advanced controls -->
-          <div
-            class="control-section"
-            v-if="visibleControls.some((c) => c.id === 'trellis-field') || true"
-          >
+          <div class="control-section" v-if="visibleControls.some((c) => c.id === 'trellis-field') || true">
             <label class="control-section-label">Advanced</label>
             <!-- Debug mode toggle -->
             <div class="control-group no-drag">
               <label class="chart-label" for="debug-mode-toggle">Debug Mode</label>
               <div class="toggle-switch-container">
                 <label class="toggle-switch">
-                  <input
-                    type="checkbox"
-                    id="debug-mode-toggle"
-                    :checked="internalConfig.showDebug"
-                    @change="toggleDebugMode"
-                    data-testid="debug-mode-toggle"
-                  />
+                  <input type="checkbox" id="debug-mode-toggle" :checked="internalConfig.showDebug"
+                    @change="toggleDebugMode" data-testid="debug-mode-toggle" />
                   <span class="toggle-slider"></span>
                 </label>
               </div>
             </div>
             <!-- Existing trellis field control -->
-            <div
-              v-for="control in visibleControls.filter((c) => c.id === 'trellis-field')"
-              :key="control.id"
-              class="control-group no-drag"
-            >
+            <div v-for="control in visibleControls.filter((c) => c.id === 'trellis-field')" :key="control.id"
+              class="control-group no-drag">
               <label class="chart-label" :for="control.id">{{ control.label }}</label>
-              <select
-                :id="control.id"
-                :value="internalConfig[control.field]"
+              <select :id="control.id" :value="internalConfig[control.field]"
                 @change="updateConfig(control.field, ($event.target as HTMLInputElement).value)"
-                class="form-select no-drag"
-              >
+                class="form-select no-drag">
                 <option v-if="control.allowEmpty" value="">None</option>
-                <option
-                  v-for="column in filteredColumnsInternal(control.columnFilter)"
-                  :key="column.name"
-                  :value="column.name"
-                >
+                <option v-for="column in filteredColumnsInternal(control.columnFilter)" :key="column.name"
+                  :value="column.name">
                   {{ column.name }}{{ column.description ? ` - ${column.description}` : '' }}
                 </option>
               </select>
@@ -174,13 +119,15 @@ import Tooltip from './Tooltip.vue'
 import type { UserSettingsStoreType } from '../stores/userSettingsStore'
 import { Controls, Charts, type ChartControl } from '../dashboards/constants'
 import {
-  generateVegaSpec,
+
   determineDefaultConfig,
   filteredColumns,
   determineEligibleChartTypes,
   getGeoTraitType,
   convertTimestampToISODate,
 } from '../dashboards/helpers'
+
+import { generateVegaSpec, } from '../dashboards/spec'
 import { debounce } from '../utility/debounce'
 
 import type { ScenegraphEvent, SignalValue } from 'vega'
@@ -208,11 +155,11 @@ export default defineComponent({
     containerWidth: Number,
     onChartConfigChange: {
       type: Function as PropType<(config: ChartConfig) => void>,
-      default: () => {},
+      default: () => { },
     },
     chartSelection: {
       type: Array as PropType<Object[]>,
-      default: () => {},
+      default: () => { },
     },
   },
 
@@ -791,11 +738,11 @@ export default defineComponent({
   border-radius: 50%;
 }
 
-input:checked + .toggle-slider {
+input:checked+.toggle-slider {
   background-color: var(--special-text);
 }
 
-input:checked + .toggle-slider:before {
+input:checked+.toggle-slider:before {
   transform: translateX(16px);
 }
