+++ lib/dashboards/helpers.ts
@@ -1,51 +1,37 @@
 import { type Row, type ResultColumn } from '../editors/results'
 import { type ChartConfig } from '../editors/results'
 import { ColumnType } from '../editors/results'
-import { toRaw } from 'vue'
-import { lookupCountry } from './countryLookup'
 import { Charts } from './constants'
 
 const temporalTraits = ['year', 'month', 'day', 'hour', 'minute', 'second']
 
 const categoricalTraits = ['year', 'month', 'day', 'hour', 'minute', 'second']
 
-/**
- * Determines if more than 80% of data rows are within the US bounding box
- * @param rows - Array of data objects
- * @param latField - Name of the field containing latitude value
- * @param longField - Name of the field containing longitude value
- * @returns Boolean indicating if more than 80% of points are in the US bounding box
- */
-function isDataMostlyInUS<T>(rows: T[], latField: keyof T, longField: keyof T): boolean {
-  // US bounding box coordinates
-  const top = 49.3457868 // north lat
-  const left = -124.7844079 // west long
-  const right = -66.9513812 // east long
-  const bottom = 24.7433195 // south lat
-
-  // Count points inside the bounding box
-  let insideCount = 0
-
-  for (const row of rows) {
-    const lat = Number(row[latField])
-    const lng = Number(row[longField])
-
-    // Check if point is inside the bounding box
-    if (bottom <= lat && lat <= top && left <= lng && lng <= right) {
-      insideCount++
-    }
-  }
-
-  // Calculate percentage of points inside the box
-  const percentageInside = (insideCount / rows.length) * 100
-
-  // Return true if more than 80% of points are inside the box
-  return percentageInside > 80
-}
 
 export function convertTimestampToISODate(timestamp: number): Date {
   const date = new Date(timestamp)
   return date
+}
+
+
+export const getColumnHasTrait = (
+  fieldName: string,
+  columns: Map<string, ResultColumn>,
+  trait: string,
+): boolean => {
+  if (!fieldName || !columns.get(fieldName)) return false
+
+  const column = columns.get(fieldName)
+  if (!column) return false
+  return getColumnHasTraitInternal(column, trait)
+}
+
+
+const getColumnHasTraitInternal = (column: ResultColumn, trait: string) => {
+  if (column.traits && column.traits.some((t) => t.endsWith(trait))) {
+    return true
+  }
+  return false
 }
 
 // Helper functions to identify column types
@@ -105,105 +91,8 @@ export const isCategoricalColumn = (column: ResultColumn): boolean => {
   return false
 }
 
-const getVegaFieldType = (fieldName: string, columns: Map<string, ResultColumn>): string => {
-  if (!fieldName || !columns.get(fieldName)) return 'nominal'
 
-  const column = columns.get(fieldName)
-  if (!column) return 'nominal'
-  if (isTemporalColumn(column)) {
-    if ([ColumnType.DATE, ColumnType.DATETIME, ColumnType.TIMESTAMP].includes(column.type)) {
-      return 'temporal'
-    }
-    return 'ordinal'
-  } else if (isNumericColumn(column)) {
-    return 'quantitative'
-  } else {
-    return 'nominal'
-  }
-}
 
-const getColumnHasTraitInternal = (column: ResultColumn, trait: string) => {
-  if (column.traits && column.traits.some((t) => t.endsWith(trait))) {
-    return true
-  }
-  return false
-}
-
-const getColumnHasTrait = (
-  fieldName: string,
-  columns: Map<string, ResultColumn>,
-  trait: string,
-): boolean => {
-  if (!fieldName || !columns.get(fieldName)) return false
-
-  const column = columns.get(fieldName)
-  if (!column) return false
-  return getColumnHasTraitInternal(column, trait)
-}
-
-const getFormatHint = (fieldName: string, columns: Map<string, ResultColumn>): any => {
-  if (!fieldName || !columns.get(fieldName)) return {}
-
-  const column = columns.get(fieldName)
-  if (!column) return {}
-
-  switch (column.type) {
-    case ColumnType.MONEY:
-      return { format: '$,.2f' }
-    case ColumnType.PERCENT:
-      return { format: '.1%' }
-    case ColumnType.DATE:
-      return { timeUnit: 'yearmonthdate' }
-    case ColumnType.TIME:
-      return { timeUnit: 'hoursminutesseconds' }
-    case ColumnType.DATETIME:
-      return { timeUnit: 'yearmonthdate-hours' }
-    case ColumnType.INTEGER:
-      return {}
-    default:
-      return {}
-  }
-}
-
-// Generate tooltip fields with proper formatting
-const generateTooltipFields = (
-  _: string,
-  xField: string,
-  yField: string,
-  columns: Map<string, ResultColumn>,
-  colorField?: string,
-): any[] => {
-  const fields: any[] = []
-
-  if (xField && columns.get(xField)) {
-    fields.push({
-      field: xField,
-      type: getVegaFieldType(xField, columns),
-      title: columns.get(xField)?.description || xField,
-      ...getFormatHint(xField, columns),
-    })
-  }
-
-  if (yField && columns.get(yField)) {
-    fields.push({
-      field: yField,
-      type: getVegaFieldType(yField, columns),
-      title: columns.get(yField)?.description || yField,
-      ...getFormatHint(yField, columns),
-    })
-  }
-
-  if (colorField && columns.get(colorField)) {
-    fields.push({
-      field: colorField,
-      type: getVegaFieldType(colorField, columns),
-      title: columns.get(colorField)?.description || colorField,
-      ...getFormatHint(colorField, columns),
-    })
-  }
-
-  return fields
-}
 
 // Interface for selection configuration
 export interface SelectionConfig {
@@ -216,796 +105,7 @@ export interface SelectionConfig {
   resolve?: 'global' | 'union' | 'intersect'
 }
 
-export const generateVegaSpec = (
-  data: readonly Row[] | null,
-  config: ChartConfig,
-  columns: Map<string, ResultColumn>,
-  chartSelection: Object[] | null,
-  isMobile: boolean = false,
-) => {
-  let intChart: Array<Partial<ChartConfig>> = chartSelection
-    ? chartSelection.map((x) => toRaw(x))
-    : []
-  let spec: any = {
-    $schema: 'https://vega.github.io/schema/vega-lite/v6.json',
-    data: { values: data },
-    width: 'container',
-    // 28 is the chart control height
-    height: 'container',
-    // height: isMobile ? containerHeight : containerHeight ? containerHeight - 150 : 'container',
-    params: [
-      {
-        name: 'highlight',
-        select: {
-          type: 'point',
-          on: 'pointerover',
-          clear: 'mouseout',
-        },
-      },
-      {
-        name: 'select',
-        select: {
-          type: 'point',
-          on: 'click,touchend',
-        },
-        value: intChart,
-        nearest: true,
-      },
-    ],
-    mark: {
-      type: 'bar',
-      fill: '#4C78A8',
-      stroke: 'black',
-      cursor: 'pointer',
-    },
-    config: {
-      scale: {
-        bandPaddingInner: 0.2,
-      },
-    },
-  }
 
-  // Basic encoding object that we'll modify based on chart type
-  let encoding: any = {}
-  let legendConfig = {}
-  if (isMobile) {
-    legendConfig = {
-      legend: {
-        orient: 'bottom',
-        direction: 'horizontal',
-      },
-    }
-  }
-
-  // Add color encoding if specified (and not for special chart types)
-  if (config.colorField && !['heatmap'].includes(config.chartType)) {
-    const fieldType = getVegaFieldType(config.colorField, columns)
-    encoding.color = {
-      field: config.colorField,
-      type: fieldType,
-      title: columns.get(config.colorField)?.description || config.colorField,
-      scale: fieldType === 'quantitative' ? { scheme: 'viridis' } : { scheme: 'category10' },
-      ...getFormatHint(config.colorField, columns),
-      ...legendConfig,
-    }
-  } else {
-    encoding.color = {
-      ...legendConfig,
-    }
-  }
-
-  // Handle trellis (facet) layout if specified
-  if (config.trellisField && config.chartType === 'line') {
-    spec.facet = {
-      field: config.trellisField,
-      type: getVegaFieldType(config.trellisField, columns),
-      title: columns.get(config.trellisField)?.description || config.trellisField,
-    }
-    spec.spec = { width: 'container', height: 200 }
-  }
-
-  const tooltipFields = generateTooltipFields(
-    config.chartType,
-    config.xField || '',
-    config.yField || '',
-    columns,
-    config.colorField,
-  )
-
-  // Build encodings for specific chart types
-  switch (config.chartType) {
-    case 'bar':
-      const barSpec = {
-        mark: 'bar',
-        encoding: {
-          x: {
-            field: config.xField,
-            type: getVegaFieldType(config.xField || '', columns),
-            title: columns.get(config.xField || '')?.description || config.xField,
-            axis: { labelAngle: -45 },
-            ...getFormatHint(config.xField || '', columns),
-          },
-          y: {
-            field: config.yField,
-            type: getVegaFieldType(config.yField || '', columns),
-            title: columns.get(config.yField || '')?.description || config.yField,
-            ...getFormatHint(config.yField || '', columns),
-          },
-          fillOpacity: {
-            condition: { param: 'select', value: 1 },
-            value: 0.3,
-          },
-          strokeWidth: {
-            condition: [
-              {
-                param: 'select',
-                empty: false,
-                value: 2,
-              },
-              {
-                param: 'highlight',
-                empty: false,
-                value: 1,
-              },
-            ],
-            value: 0,
-          },
-          tooltip: tooltipFields,
-          ...encoding,
-        },
-      }
-
-      if (config.trellisField) {
-        spec.spec = barSpec
-      } else {
-        spec = { ...spec, ...barSpec }
-      }
-
-      break
-
-    case 'barh':
-      const barHSpec = {
-        mark: 'bar',
-        encoding: {
-          y: {
-            field: config.yField,
-            type: getVegaFieldType(config.yField || '', columns),
-            title: columns.get(config.yField || '')?.description || config.yField,
-            sort: '-x',
-            axis: {
-              labelExpr: isMobile
-                ? "datum.label.length > 13 ? slice(datum.label, 0, 10) + '...' : datum.label"
-                : 'datum.label',
-            },
-            ...getFormatHint(config.yField || '', columns),
-          },
-          x: {
-            field: config.xField,
-            type: getVegaFieldType(config.xField || '', columns),
-            title: columns.get(config.xField || '')?.description || config.xField,
-            ...getFormatHint(config.xField || '', columns),
-          },
-          fillOpacity: {
-            condition: { param: 'select', value: 1 },
-            value: 0.3,
-          },
-          tooltip: tooltipFields,
-          ...encoding,
-        },
-      }
-
-      if (config.trellisField) {
-        spec.spec = barHSpec
-      } else {
-        spec = { ...spec, ...barHSpec }
-      }
-      break
-
-    case 'line':
-      const lineSpec = {
-        data: undefined,
-        params: [],
-        layer: [
-          {
-            params: [
-              {
-                name: 'brush',
-                select: { type: 'interval', encodings: ['x'], value: intChart },
-                //@ts-ignore
-                value:
-                  intChart.length > 0 && config.xField
-                    ? { x: intChart[0][config.xField as keyof typeof config] }
-                    : undefined,
-              },
-            ],
-            mark: { type: 'line', color: 'lightgray' },
-            data: { values: data },
-            encoding: {
-              x: {
-                field: config.xField,
-                type: getVegaFieldType(config.xField || '', columns),
-                title: columns.get(config.xField || '')?.description || config.xField,
-                ...getFormatHint(config.xField || '', columns),
-              },
-              y: {
-                field: config.yField,
-                type: getVegaFieldType(config.yField || '', columns),
-                title: columns.get(config.yField || '')?.description || config.yField,
-                ...getFormatHint(config.yField || '', columns),
-              },
-              tooltip: tooltipFields,
-              ...encoding,
-            },
-          },
-          {
-            transform: [
-              {
-                filter: {
-                  param: 'brush',
-                },
-              },
-            ],
-            mark: { type: 'line', color: 'steelblue' },
-            data: { values: data },
-            encoding: {
-              x: {
-                field: config.xField,
-                type: getVegaFieldType(config.xField || '', columns),
-                title: columns.get(config.xField || '')?.description || config.xField,
-                ...getFormatHint(config.xField || '', columns),
-              },
-              y: {
-                field: config.yField,
-                type: getVegaFieldType(config.yField || '', columns),
-                title: columns.get(config.yField || '')?.description || config.yField,
-                ...getFormatHint(config.yField || '', columns),
-              },
-              tooltip: tooltipFields,
-              ...encoding,
-            },
-          },
-        ],
-      }
-
-      spec = { ...spec, ...lineSpec }
-      break
-
-    case 'area':
-      const areaSpec = {
-        data: undefined,
-        // params: [],
-        layer: [
-          {
-            params: [
-              {
-                name: 'brush',
-                select: { type: 'interval', encodings: ['x'], value: intChart },
-                //@ts-ignore
-                value:
-                  intChart.length > 0 && config.xField
-                    ? { x: intChart[0][config.xField as keyof typeof config] }
-                    : undefined,
-              },
-            ],
-            mark: { type: 'area', line: true },
-            data: { values: data },
-            encoding: {
-              x: {
-                field: config.xField,
-                type: getVegaFieldType(config.xField || '', columns),
-                title: columns.get(config.xField || '')?.description || config.xField,
-                ...getFormatHint(config.xField || '', columns),
-              },
-              y: {
-                field: config.yField,
-                type: getVegaFieldType(config.yField || '', columns),
-                title: columns.get(config.yField || '')?.description || config.yField,
-                ...getFormatHint(config.yField || '', columns),
-              },
-              tooltip: tooltipFields,
-              ...encoding,
-            },
-          },
-          {
-            transform: [
-              {
-                filter: {
-                  param: 'brush',
-                },
-              },
-            ],
-            mark: { type: 'area', line: true },
-            data: { values: data },
-            encoding: {
-              x: {
-                field: config.xField,
-                type: getVegaFieldType(config.xField || '', columns),
-                title: columns.get(config.xField || '')?.description || config.xField,
-                ...getFormatHint(config.xField || '', columns),
-              },
-              y: {
-                field: config.yField,
-                type: getVegaFieldType(config.yField || '', columns),
-                title: columns.get(config.yField || '')?.description || config.yField,
-                ...getFormatHint(config.yField || '', columns),
-              },
-              tooltip: tooltipFields,
-              ...encoding,
-            },
-          },
-        ],
-      }
-
-      spec = { ...spec, ...areaSpec }
-      console.log(spec)
-      break
-
-    case 'point':
-      const pointSpec = {
-        mark: { type: 'point', filled: true },
-        encoding: {
-          x: {
-            field: config.xField,
-            type: getVegaFieldType(config.xField || '', columns),
-            title: columns.get(config.xField || '')?.description || config.xField,
-            ...getFormatHint(config.xField || '', columns),
-          },
-          y: {
-            field: config.yField,
-            type: getVegaFieldType(config.yField || '', columns),
-            title: columns.get(config.yField || '')?.description || config.yField,
-            ...getFormatHint(config.yField || '', columns),
-          },
-          tooltip: tooltipFields,
-          ...encoding,
-        },
-      }
-
-      if (config.trellisField) {
-        spec.spec = pointSpec
-      } else {
-        spec = { ...spec, ...pointSpec }
-      }
-      break
-
-    case 'headline':
-      const headlineSpec = {
-        $schema: 'https://vega.github.io/schema/vega-lite/v6.json',
-        description: 'A simple headline metric display',
-        width: 'container',
-        height: 'container',
-        data: {
-          values: data ? data[0] : [],
-        },
-        layer: [
-          {
-            mark: {
-              type: 'text',
-              fontSize: 36,
-              fontWeight: 'bold',
-              align: 'center',
-              baseline: 'middle',
-              dx: 0,
-              dy: -20,
-            },
-            encoding: {
-              text: { field: config.xField, type: 'quantitative' },
-              color: { value: '#333333' },
-            },
-          },
-          {
-            mark: {
-              type: 'text',
-              fontSize: 14,
-              fontWeight: 'normal',
-              align: 'center',
-              baseline: 'top',
-              dx: 0,
-              dy: 10,
-            },
-            encoding: {
-              text: { value: config.xField },
-              color: { value: '#666666' },
-            },
-          },
-        ],
-        config: {
-          view: { stroke: null },
-          axis: { grid: false, domain: false },
-        },
-      }
-
-      spec = { ...headlineSpec }
-      break
-    case 'heatmap':
-      const heatmapSpec = {
-        mark: 'rect',
-        encoding: {
-          x: {
-            field: config.xField,
-            type: getVegaFieldType(config.xField || '', columns),
-            title: columns.get(config.xField || '')?.description || config.xField,
-            ...getFormatHint(config.xField || '', columns),
-          },
-          y: {
-            field: config.yField,
-            type: getVegaFieldType(config.yField || '', columns),
-            title: columns.get(config.yField || '')?.description || config.yField,
-            ...getFormatHint(config.yField || '', columns),
-          },
-          color: {
-            field: config.colorField,
-            type: getVegaFieldType(config.colorField || '', columns),
-            title: columns.get(config.colorField || '')?.description || config.colorField,
-            scale: { scheme: 'viridis' },
-            ...getFormatHint(config.colorField || '', columns),
-          },
-          tooltip: tooltipFields,
-        },
-      }
-
-      if (config.trellisField) {
-        spec.spec = heatmapSpec
-      } else {
-        spec = { ...spec, ...heatmapSpec }
-      }
-      break
-    case 'usa-map':
-      // First, we need to set up the correct projection
-
-      if (config.xField && config.yField) {
-        //@ts-ignore
-        if (isDataMostlyInUS(data, config.yField, config.xField)) {
-          spec.projection = {
-            type: 'albersUsa',
-          }
-        }
-        const usaMapSpec = {
-          params: [
-            {
-              name: 'highlight',
-              select: {
-                type: 'point',
-                on: 'pointerover',
-                clear: 'mouseout',
-              },
-            },
-            {
-              name: 'select',
-              select: 'point',
-              // override the default selection for now
-            },
-          ],
-          layer: [
-            {
-              data: {
-                url: 'https://cdn.jsdelivr.net/npm/vega-datasets@2.2.0/data/us-10m.json',
-                format: { type: 'topojson', feature: 'states' },
-              },
-              mark: { type: 'geoshape', fill: '#e5e5e5', stroke: 'white' },
-            },
-            {
-              mark: { type: 'circle', tooltip: true },
-              encoding: {
-                longitude: { field: config.xField, type: 'quantitative' },
-                latitude: { field: config.yField, type: 'quantitative' },
-                size: {
-                  field: config.sizeField,
-                  type: 'quantitative',
-                  title: config.sizeField,
-                  scale: {
-                    type: 'quantize',
-                    nice: true,
-                  },
-                },
-
-                color: config.colorField
-                  ? {
-                      field: config.colorField,
-                      type: 'quantitative',
-                      title: config.colorField,
-                      scale: { scheme: 'viridis' },
-                      ...legendConfig,
-                    }
-                  : { value: 'steelblue' },
-                tooltip: [
-                  {
-                    field: config.xField,
-                    type: 'quantitative',
-                    title: config.xField
-                      ? columns.get(config.xField)?.description || config.xField
-                      : config.xField,
-                  },
-                  {
-                    field: config.yField,
-                    type: 'quantitative',
-                    title: config.yField
-                      ? columns.get(config.yField)?.description || config.yField
-                      : config.yField,
-                  },
-                  {
-                    field: config.sizeField,
-                    type: 'quantitative',
-                    title: config.sizeField
-                      ? columns.get(config.sizeField)?.description || config.sizeField
-                      : config.sizeField,
-                  },
-                ],
-              },
-            },
-          ],
-        }
-
-        // Replace the existing spec with the updated usaMapSpec
-        spec = { ...spec, ...usaMapSpec }
-      } else if (config.geoField && getColumnHasTrait(config.geoField, columns, 'us_state_short')) {
-        const base = {
-          $schema: 'https://vega.github.io/schema/vega-lite/v6.json',
-          width: 'container',
-          height: 'container',
-          // config: { scale: { bandPaddingInner: 0.2 }, view: { stroke: null } },
-          projection: { type: 'albersUsa' },
-          layer: [
-            {
-              data: {
-                url: 'https://cdn.jsdelivr.net/npm/vega-datasets@2/data/us-10m.json',
-                format: { type: 'topojson', feature: 'states' },
-              },
-              mark: { type: 'geoshape', fill: '#e5e5e5', stroke: 'white' },
-            },
-            {
-              data: {
-                url: 'https://cdn.jsdelivr.net/npm/vega-datasets@2/data/us-10m.json',
-                format: { type: 'topojson', feature: 'states' },
-              },
-              mark: { type: 'geoshape' },
-              params: [
-                {
-                  name: 'highlight',
-                  select: { type: 'point', on: 'mouseover', clear: 'mouseout' },
-                },
-                { name: 'select', select: 'point', value: [] },
-              ],
-              encoding: {
-                color: {
-                  field: config.colorField,
-                  type: 'quantitative',
-                  scale: { type: 'quantize', nice: true, zero: true },
-                  legend: { title: config.colorField },
-                },
-                opacity: { condition: { param: 'select', value: 1 }, value: 0.3 },
-                stroke: {
-                  condition: { param: 'highlight', empty: false, value: 'black' },
-                  value: null,
-                },
-                strokeWidth: {
-                  condition: { param: 'highlight', empty: false, value: 2 },
-                  value: 0.5,
-                },
-              },
-              transform: [
-                {
-                  lookup: 'id',
-                  from: {
-                    data: {
-                      values: [
-                        { id: 1, abbr: 'AL' },
-                        { id: 2, abbr: 'AK' },
-                        { id: 4, abbr: 'AZ' },
-                        { id: 5, abbr: 'AR' },
-                        { id: 6, abbr: 'CA' },
-                        { id: 8, abbr: 'CO' },
-                        { id: 9, abbr: 'CT' },
-                        { id: 10, abbr: 'DE' },
-                        { id: 11, abbr: 'DC' },
-                        { id: 12, abbr: 'FL' },
-                        { id: 13, abbr: 'GA' },
-                        { id: 15, abbr: 'HI' },
-                        { id: 16, abbr: 'ID' },
-                        { id: 17, abbr: 'IL' },
-                        { id: 18, abbr: 'IN' },
-                        { id: 19, abbr: 'IA' },
-                        { id: 20, abbr: 'KS' },
-                        { id: 21, abbr: 'KY' },
-                        { id: 22, abbr: 'LA' },
-                        { id: 23, abbr: 'ME' },
-                        { id: 24, abbr: 'MD' },
-                        { id: 25, abbr: 'MA' },
-                        { id: 26, abbr: 'MI' },
-                        { id: 27, abbr: 'MN' },
-                        { id: 28, abbr: 'MS' },
-                        { id: 29, abbr: 'MO' },
-                        { id: 30, abbr: 'MT' },
-                        { id: 31, abbr: 'NE' },
-                        { id: 32, abbr: 'NV' },
-                        { id: 33, abbr: 'NH' },
-                        { id: 34, abbr: 'NJ' },
-                        { id: 35, abbr: 'NM' },
-                        { id: 36, abbr: 'NY' },
-                        { id: 37, abbr: 'NC' },
-                        { id: 38, abbr: 'ND' },
-                        { id: 39, abbr: 'OH' },
-                        { id: 40, abbr: 'OK' },
-                        { id: 41, abbr: 'OR' },
-                        { id: 42, abbr: 'PA' },
-                        { id: 44, abbr: 'RI' },
-                        { id: 45, abbr: 'SC' },
-                        { id: 46, abbr: 'SD' },
-                        { id: 47, abbr: 'TN' },
-                        { id: 48, abbr: 'TX' },
-                        { id: 49, abbr: 'UT' },
-                        { id: 50, abbr: 'VT' },
-                        { id: 51, abbr: 'VA' },
-                        { id: 53, abbr: 'WA' },
-                        { id: 54, abbr: 'WV' },
-                        { id: 55, abbr: 'WI' },
-                        { id: 56, abbr: 'WY' },
-                      ],
-                    },
-                    key: 'id',
-                    fields: ['abbr'],
-                  },
-                },
-                {
-                  lookup: 'abbr',
-                  from: {
-                    data: {
-                      values: data,
-                    },
-                    key: config.geoField,
-                    fields: [config.colorField, config.sizeField, config.geoField].filter((x) => x),
-                  },
-                },
-              ],
-            },
-          ],
-        }
-        spec = {
-          ...base,
-        }
-      } else if (config.geoField && getColumnHasTrait(config.geoField, columns, 'country')) {
-        let lookupField: string = config.geoField
-        //@ts-ignore
-        const idLookup = (row) => {
-          const lookup = lookupCountry(row[lookupField], 'name')
-          if (lookup) {
-            return Number(lookup['country-code'])
-          } else {
-            return null
-          }
-        }
-        const base = {
-          $schema: 'https://vega.github.io/schema/vega-lite/v6.json',
-          width: 'container',
-          height: 'container',
-          layer: [
-            {
-              data: {
-                url: 'https://cdn.jsdelivr.net/npm/vega-datasets@2/data/world-110m.json',
-                format: { type: 'topojson', feature: 'countries' },
-              },
-              transform: [
-                {
-                  filter: 'datum.id !== 10', // Filter out Antarctica (country code 10)
-                },
-              ],
-              mark: { type: 'geoshape', fill: '#e0e0e0', stroke: '#bbb', strokeWidth: 0.5 },
-              projection: { type: 'mercator' },
-            },
-            {
-              data: {
-                url: 'https://cdn.jsdelivr.net/npm/vega-datasets@2/data/world-110m.json',
-                format: { type: 'topojson', feature: 'countries' },
-              },
-              transform: [
-                {
-                  lookup: 'id',
-                  from: {
-                    data: {
-                      values: data?.map((row) => {
-                        return {
-                          id: idLookup(row),
-                          ...row,
-                        }
-                      }),
-                    },
-                    key: 'id',
-                    fields: [config.colorField, config.sizeField, config.geoField].filter((x) => x),
-                  },
-                },
-              ],
-              params: [
-                {
-                  name: 'highlight',
-                  select: { type: 'point', on: 'mouseover', clear: 'mouseout' },
-                },
-                {
-                  name: 'select',
-                  select: 'point',
-                  value: intChart,
-                },
-              ],
-              mark: { type: 'geoshape' },
-              projection: { type: 'mercator' },
-              encoding: {
-                color: {
-                  field: config.colorField,
-                  type: 'quantitative',
-                  scale: { type: 'quantize', nice: true },
-                  legend: { title: config.colorField },
-                },
-                opacity: { condition: { param: 'select', value: 1 }, value: 0.3 },
-                tooltip: [
-                  { field: config.geoField, type: 'nominal', title: config.geoField },
-                  { field: config.colorField, type: 'quantitative', title: config.colorField },
-                ],
-                stroke: {
-                  condition: { param: 'highlight', empty: false, value: 'black' },
-                  value: '#666',
-                },
-                strokeWidth: {
-                  condition: { param: 'highlight', empty: false, value: 2 },
-                  value: 0.5,
-                },
-              },
-            },
-          ],
-          config: { view: { stroke: null } },
-        }
-        spec = {
-          ...base,
-        }
-      }
-
-      break
-    case 'boxplot':
-      const boxplotSpec = {
-        mark: { type: 'boxplot', extent: 'min-max' },
-        encoding: {
-          x: {
-            field: config.groupField,
-            type: getVegaFieldType(config.groupField || '', columns),
-            title: columns.get(config.groupField || '')?.description || config.groupField,
-            ...getFormatHint(config.groupField || '', columns),
-          },
-          y: {
-            field: config.yField,
-            type: getVegaFieldType(config.yField || '', columns),
-            title: columns.get(config.yField || '')?.description || config.yField,
-            ...getFormatHint(config.yField || '', columns),
-          },
-          tooltip: tooltipFields,
-          ...encoding,
-        },
-      }
-
-      if (config.trellisField) {
-        spec.spec = boxplotSpec
-      } else {
-        spec = { ...spec, ...boxplotSpec }
-      }
-      break
-  }
-
-  // universal configs
-  if (isMobile) {
-    spec.point = {
-      size: 80, // Larger hit area for touch targets
-    }
-    spec.signals = [
-      {
-        name: 'touchSignal',
-        on: [
-          {
-            events: 'touchend',
-            update: 'datum', // This is crucial for touch events
-          },
-        ],
-      },
-    ]
-  }
-  return spec
-}
 
 export const columHasTraitEnding = (column: ResultColumn, trait: string): boolean => {
   if (column.traits) {
